pyproject.toml
[tool.poetry]
name = "cv"
version = "0.1.0"
description = ""
authors = ["powerwook <powerwook@naver.com>"]
readme = "README.md"
packages = [{include = "cv", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "*"
mediapipe = "*"
numpy = "*"
pandas = "*"
scipy = "*"
fastapi = "*"
uvicorn = "*"
python-socketio = "*"
tensorflow = ">=2.15"
scikit-learn = "^1.7.2"
streamlit = "^1.50.0"
streamlit-webrtc = "^0.63.5"
pillow = "^11.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
